# eth/tracers: avoid using blockCtx concurrently